Separate cli from api
